% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Germany}
\alias{Germany}
\title{3 variables for Germany}
\format{A data frame with 20 rows and 3 variables:
\itemize{
  \item Equipo: Team ID (character)
  \item Favor: Goals/xG Goals Favor (double)
  \item Contra: Goals/xG Goals Against (double)
}}
\usage{
Germany
}
\description{
Se ha calculado el ratio de Goles y xG a favor y en contra para los equipos espa?noles
}
\keyword{datasets}
