% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escudos.R
\name{PlotXY_Badges}
\alias{PlotXY_Badges}
\title{Esta funcion crea un grafico scatter plot con los escudos de las principales ligas europeas}
\usage{
PlotXY_Badges(df, league_code)
}
\arguments{
\item{df}{Un df}

\item{league}{Un literal que marca la liga que se quiere mostrar}
}
\value{
El grafico a partir de la tabla \code{df} con los escudos
}
\description{
Esta funcion crea un grafico scatter plot con los escudos de las principales ligas europeas
}
\examples{
PlotXY_Badges(df,"ESP")
}
