% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Teams}
\alias{Teams}
\title{3 variables for Teams}
\format{A data frame with 98 rows and 3 variables:
\itemize{
  \item Country: Country ID (character)
  \item Team: Name (character)
  \item Code: Code (character)
}}
\usage{
Teams
}
\description{
Listado de los equipos incluidos en la libreria
}
\keyword{datasets}
